feat(apollo-react): expose canvas handle positioning utility
